Se agrega el modelo cart
