+++ api/src/db.js
@@ -27,11 +27,19 @@ fs.readdirSync(path.join(__dirname, '/models'))
  let capsEntries = entries.map((entry) => [entry[0][0].toUpperCase()+entry[0].slice(1), entry[1]]);
  sequelize.models = Object.fromEntries(capsEntries);
 
- const { Product , Category, User} = sequelize.models;
+ const { Product , Category, User, Cart} = sequelize.models;
+
+ //Relaciones entre usuarios y carts
+ User.hasOne(Cart, {as:'cartUser', foreignKey:"cartUserId"})
+ Cart.belongsTo(User, {as: 'cartUser'} )
 
  //Relaciones entre usuarios y productos
 User.hasMany(Product, {as: "productsOwner", foreignKey: "ownerId"})
 Product.belongsTo(User, {as: "owner"})
+
+//Relaciones entre productos y cart
+Product.belongsToMany(Cart, {through: "cart_product"})
+Cart.belongsToMany(Product, {through: "cart_product"})
 
  //Relaciones entre productos y categorias
  Product.belongsToMany(Category, {through: "product_category"})
@@ -41,5 +49,6 @@ Product.belongsTo(User, {as: "owner"})
     Product,
     Category,
     User,
+    Cart,
     conn: sequelize,
  }
