+++ api/src/models/Cart.js
@@ -1,0 +1,16 @@
+const { DataTypes, Model } = require("sequelize");
+
+class Cart extends Model {}
+
+module.exports = (sequelize) => {
+  return Cart.init(
+    {
+      id: {
+        type: DataTypes.INTEGER,
+        primaryKey: true,
+        autoIncrement: true
+      },
+    },
+    { sequelize, modelName: "cart", timestamps: false }
+  );
+};
